RepText/requirements.txt
torch
transformers
accelerate
safetensors
huggingface-hub
librosa
soundfile
pydub
fastapi
uvicorn
firebase-admin
python-dotenv
requests
pydantic
numpy
pandas
aiosqlite
opencv-python-headless
google-cloud-storage

# Training dependencies
diffusers>=0.27.0
pillow
pyyaml
tqdm
wandb
torchvision
datasets
bitsandbytes  # For 8-bit optimizers (optional)